pyproject.toml
[project]
name = "train-bot"
version = "0.1.0"
description = "Telegram bot that queries UK rail departures via RealTimeTrains with TransportAPI fallback"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.3",
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
]

[project.scripts]
train-bot = "train_bot.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = []
